refactor : 거리요금 enum의 메소드 클래스로 분리
